Refactoring suggestions (#111)
* refactor: replace NodeWebRtcEndpoint with composition

* refactor: rm duplicated code segment

* refactor: NodeWebRtcConnection remove duplicate .bind

* refactor: inheritance round #1

- move public method `connect` to base class, add abstract protected
  `doConnect` method as replacement for subclasses
- move public method `close` to base class, add abstract protected
  `doClose` method as replacement for subclasses
- make `connectionTimeoutRef` field private and move all related logic to base class
- make ìsFinished` field private
- add method `restartConnectionTimeout` to base class to allow a subclass to
  forcefully restart a connection timer

* refactor: add JSDOC to method doSendMessage

* refactor: move last and lastGatheringState to subclass

* refactor: add methods to base class to emit

* refactor: make paused field private in base class

- Field `paused` is private in base class
- Add protected method emitLowBackPressure thru which subclass can
  inform base calss of low back pressure
- Therefore backpressure managment is now taken care of by base class
- Made `setFlushRef` private in base class

* refactor: WebRtcConnection#emitMessage for handling message

* refactor: fix accidental lastGatheringState private

* refactor: separate logger for base and sub classes

* refactor: privatize WebRtcConnection fields

* fix: add class-level JSDoc to WebRtcConnection

* fix: make eslint happy

* fix: address PR comment, move factories
