+++ packages/network/src/connection/WebRtcConnection.ts
@@ -81,38 +81,59 @@ export function isOffering(myId: string, theirId: string): boolean {
     return myId < theirId
 }
 
+/**
+ * Shared base class for WebRTC connections implemented in different libraries.
+ * Encapsulates the common needs of such connections such as:
+ *
+ *  - Determining offerer / answerer roles upon connecting
+ *  - Connection timeout
+ *  - Message queueing and retries on message delivery failures
+ *  - Backpressure handling
+ *  - Ping/Pong mechanism for RTT calculation and dead connection detection
+ *  - Deferred promise handling in case of connection re-attempts
+ *  - Closing of connection and associated clean up
+ *  - Ensuring event loop isn't greedily reserved for long periods of time
+ *
+ *  Implementers of this base class should make sure to implement the
+ *  abstract methods. Implementers should also make sure their base classes
+ *  invoke all "emit"-prefixed protected methods:
+ *  - emitOpen
+ *  - emitLocalDescription
+ *  - emitLocalCandidate
+ *  - emitMessage
+ *  - emitLowBackpressure
+ *
+ *  See the respective JSDocs for more information.
+ *
+ */
 export abstract class WebRtcConnection extends ConnectionEmitter {
-    public readonly id: string
-    private connectionId = 'none'
-    private peerInfo: PeerInfo
     private readonly routerId: string
-    
     private readonly maxPingPongAttempts: number
     private readonly pingInterval: number
     private readonly flushRetryTimeout: number
     private readonly messageQueue: MessageQueue<string>
-   
+    private readonly baseLogger: Logger
+
+    private connectionId = 'none'
+    private peerInfo: PeerInfo
     private flushRef: NodeJS.Immediate | null
     private flushTimeoutRef: NodeJS.Timeout | null
-
+    private connectionTimeoutRef: NodeJS.Timeout | null
     private pingTimeoutRef: NodeJS.Timeout | null
     private deferredConnectionAttempt: DeferredConnectionAttempt | null
+    private readonly newConnectionTimeout: number
+    private paused: boolean
+    private isFinished: boolean
+    private pingAttempts = 0
+    private rtt: number | null
+    private rttStart: number | null
 
-    protected readonly logger: Logger
+    protected readonly id: string
     protected readonly maxMessageSize: number
-    protected isFinished: boolean
     protected readonly selfId: string
-    protected readonly stunUrls: string[]
-    protected connectionTimeoutRef: NodeJS.Timeout | null
-    protected readonly newConnectionTimeout: number 
-    protected lastState: string | null
-    protected lastGatheringState: string | null
-    protected paused: boolean
+    protected readonly stunUrls: ReadonlyArray<string>
     protected readonly bufferThresholdHigh: number
     protected readonly bufferThresholdLow: number
-    protected pingAttempts = 0
-    protected rtt: number | null
-    protected rttStart: number | null
 
     constructor({
         selfId,
@@ -146,12 +167,9 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
         this.flushRetryTimeout = flushRetryTimeout
         this.messageQueue = messageQueue
         this.deferredConnectionAttempt = deferredConnectionAttempt
-        this.logger = new Logger(module, `${NameDirectory.getName(this.getPeerId())}/${ID}`)
+        this.baseLogger = new Logger(module, `${NameDirectory.getName(this.getPeerId())}/${ID}`)
         this.isFinished = false
-
         this.paused = false
-        this.lastState = null
-        this.lastGatheringState = null
 
         this.flushTimeoutRef = null
         this.connectionTimeoutRef = null
@@ -161,11 +179,23 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
         this.rtt = null
         this.rttStart = null
 
-        this.logger.trace('create %o', {
+        this.baseLogger.trace('create %o', {
             selfId: this.selfId,
             messageQueue: this.messageQueue.size(),
             peerInfo: this.peerInfo,
         })
+    }
+
+    connect(): void {
+        if (this.isFinished) {
+            throw new Error('Connection already closed.')
+        }
+        this.doConnect()
+        this.connectionTimeoutRef = setTimeout(() => {
+            if (this.isFinished) { return }
+            this.baseLogger.warn(`connection timed out after ${this.newConnectionTimeout}ms`)
+            this.close(new Error(`timed out after ${this.newConnectionTimeout}ms`))
+        }, this.newConnectionTimeout)
     }
 
     getDeferredConnectionAttempt(): DeferredConnectionAttempt | null {
@@ -178,7 +208,19 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
         return att
     }
 
-    protected doClose(err?: Error): void {
+    close(err?: Error): void {
+        if (this.isFinished) {
+            // already closed, noop
+            return
+        }
+        this.isFinished = true
+
+        if (err) {
+            this.baseLogger.warn('conn.close(): %s', err)
+        } else {
+            this.baseLogger.trace('conn.close()')
+        }
+
         if (this.flushRef) {
             clearImmediate(this.flushRef)
         }
@@ -197,19 +239,17 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
         this.pingTimeoutRef = null
         this.flushRef = null
 
+        try {
+            this.doClose(err)
+        } catch(e) {
+            this.baseLogger.warn(`doClose (subclass) threw: %s`, e)
+        }
+
         if (err) {
             this.emitClose(err)
             return
         }
         this.emitClose('closed')
-    }
-    protected emitOpen(): void {
-        if (this.deferredConnectionAttempt) {
-            const def = this.deferredConnectionAttempt
-            this.deferredConnectionAttempt = null
-            def.resolve()
-        }
-        this.emit('open')
     }
 
     protected emitClose(reason: Error | string): void {
@@ -234,15 +274,6 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
         return this.messageQueue.add(message)
     }
 
-    protected setFlushRef(): void {
-        if (this.flushRef === null) {
-            this.flushRef = setImmediate(() => {
-                this.flushRef = null
-                this.attemptToFlushMessages()
-            })
-        }
-    }
-
     setPeerInfo(peerInfo: PeerInfo): void {
         this.peerInfo = peerInfo
     }
@@ -259,17 +290,13 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
         return this.rtt
     }
 
-    getLastState(): string | null {
-        return this.lastState
-    }
-
     ping(): void {
         if (this.isFinished) {
             return
         }
         if (this.isOpen()) {
             if (this.pingAttempts >= this.maxPingPongAttempts) {
-                this.logger.warn(`failed to receive any pong after ${this.maxPingPongAttempts} ping attempts, closing connection`)
+                this.baseLogger.warn(`failed to receive any pong after ${this.maxPingPongAttempts} ping attempts, closing connection`)
                 this.close(new Error('pong not received'))
             } else {
                 this.rttStart = Date.now()
@@ -278,7 +305,7 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
                         this.doSendMessage('ping')
                     }
                 } catch (e) {
-                    this.logger.warn(`failed to send ping to ${this.peerInfo.peerId} with error: ${e}`)
+                    this.baseLogger.warn(`failed to send ping to ${this.peerInfo.peerId} with error: ${e}`)
                 }
                 this.pingAttempts += 1
             }
@@ -298,7 +325,7 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
                 this.doSendMessage('pong')
             }
         } catch (e) {
-            this.logger.warn(`failed to send pong to ${this.peerInfo.peerId} with error: ${e}`)
+            this.baseLogger.warn(`failed to send pong to ${this.peerInfo.peerId} with error: ${e}`)
         }
     }
 
@@ -308,6 +335,15 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
 
     isOffering(): boolean {
         return isOffering(this.selfId, this.peerInfo.peerId)
+    }
+
+    private setFlushRef(): void {
+        if (this.flushRef === null) {
+            this.flushRef = setImmediate(() => {
+                this.flushRef = null
+                this.attemptToFlushMessages()
+            })
+        }
     }
    
     private attemptToFlushMessages(): void {
@@ -322,7 +358,7 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
 
             const queueItem = this.messageQueue.peek()
             if (queueItem.isFailed()) {
-                this.logger.debug('popping failed queue item: %o', queueItem, numOfSuccessSends)
+                this.baseLogger.debug('popping failed queue item: %o', queueItem, numOfSuccessSends)
                 this.messageQueue.pop()
             } else if (queueItem.getMessage().length > this.getMaxMessageSize())  {
                 const errorMessage = 'Dropping message due to size '
@@ -330,7 +366,7 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
                     + ' exceeding the limit of '
                     + this.getMaxMessageSize()
                 queueItem.immediateFail(errorMessage)
-                this.logger.warn(errorMessage)
+                this.baseLogger.warn(errorMessage)
                 this.messageQueue.pop()
             } else if (this.paused || this.getBufferedAmount() >= this.bufferThresholdHigh) {
                 if (!this.paused) {
@@ -358,7 +394,7 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
                     queueItem.delivered()
                     numOfSuccessSends += 1
                 } else {
-                    this.logger.debug('queue item was not sent: %o', {
+                    this.baseLogger.debug('queue item was not sent: %o', {
                         wasOpen: isOpen,
                         numOfSuccessSends,
                         queueItem,
@@ -373,12 +409,12 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
     private processFailedMessage(queueItem: QueueItem<any>, e: Error): void {
         queueItem.incrementTries({
             error: e.toString(),
-            'connection.iceConnectionState': this.lastGatheringState,
-            'connection.connectionState': this.lastState
+            'connection.iceConnectionState': this.getLastGatheringState(),
+            'connection.connectionState': this.getLastState()
         })
         if (queueItem.isFailed()) {
             const infoText = queueItem.getErrorInfos().map((i) => JSON.stringify(i)).join('\n\t')
-            this.logger.warn('failed to send message after %d tries due to\n\t%s',
+            this.baseLogger.warn('failed to send message after %d tries due to\n\t%s',
                 MessageQueue.MAX_TRIES,
                 infoText)
             this.messageQueue.pop()
@@ -391,12 +427,89 @@ export abstract class WebRtcConnection extends ConnectionEmitter {
         }
     }
     
-    abstract connect(): void
     abstract setRemoteDescription(description: string, type: string): void
-    abstract addRemoteCandidate(candidate: string, mid: string): void 
-    abstract close(err?: Error): void
-    abstract getBufferedAmount(): number 
-    abstract getMaxMessageSize(): number 
+    abstract addRemoteCandidate(candidate: string, mid: string): void
+    abstract getBufferedAmount(): number
+    abstract getMaxMessageSize(): number
     abstract isOpen(): boolean
+    protected abstract doConnect(): void
+    protected abstract doClose(err?: Error): void
+    abstract getLastState(): string | undefined
+    abstract getLastGatheringState(): string | undefined
+
+    /**
+     * Invoked when a message is ready to be sent. Connectivity is ensured
+     * with a check to `isOpen` before invocation.
+     * @param message - mesasge to be sent
+     * @return return false if the message could not be delivered
+     */
     protected abstract doSendMessage(message: string): boolean
+
+    /**
+     * Subclass should call this method when the connection has opened.
+     */
+    protected emitOpen(): void {
+        if (this.connectionTimeoutRef !== null) {
+            clearTimeout(this.connectionTimeoutRef)
+        }
+        if (this.deferredConnectionAttempt) {
+            const def = this.deferredConnectionAttempt
+            this.deferredConnectionAttempt = null
+            def.resolve()
+        }
+        this.setFlushRef()
+        this.emit('open')
+    }
+
+    /**
+     * Subclass should call this method when a new local description is available.
+     */
+    protected emitLocalDescription(description: string, type: string): void {
+        this.emit('localDescription', type, description)
+    }
+
+    /**
+     * Subclass should call this method when a new local candidate is available.
+     */
+    protected emitLocalCandidate(candidate: string, mid: string): void {
+        this.emit('localCandidate', candidate, mid)
+    }
+
+    /**
+     * Subclass should call this method when it has received a message.
+     */
+    protected emitMessage(msg: string): void {
+        if (msg === 'ping') {
+            this.pong()
+        } else if (msg === 'pong') {
+            this.pingAttempts = 0
+            this.rtt = Date.now() - this.rttStart!
+        } else {
+            this.emit('message', msg)
+        }
+    }
+
+    /**
+     * Subclass should call this method when backpressure has reached low watermark.
+     */
+    protected emitLowBackpressure(): void {
+        if (!this.paused) {
+            return
+        }
+        this.paused = false
+        this.setFlushRef()
+        this.emit('bufferLow')
+    }
+
+    /**
+     * Forcefully restart the connection timeout (e.g. on state change) from subclass.
+     */
+    protected restartConnectionTimeout(): void {
+        clearTimeout(this.connectionTimeoutRef!)
+        this.connectionTimeoutRef = setTimeout(() => {
+            if (this.isFinished) { return }
+            this.baseLogger.warn(`connection timed out after ${this.newConnectionTimeout}ms`)
+            this.close(new Error(`timed out after ${this.newConnectionTimeout}ms`))
+        }, this.newConnectionTimeout)
+    }
 }
